+++ src/Cell.ts
@@ -14,11 +14,12 @@ import type {
     CellHoverIconMethod,
     CellStyleMethod,
     OverflowTooltipPlacement,
-    Rules,
     SpanInfo,
+    SelectorCellValueType,
 } from './types';
 import Context from './Context';
 import BaseCell from './BaseCell';
+import { Rule, Rules } from './Validator';
 export default class Cell extends BaseCell {
     formatter?: FormatterMethod;
     formatterFooter?: FormatterMethod;
@@ -48,7 +49,7 @@ export default class Cell extends BaseCell {
     render: Render;
     renderFooter: Render;
     style: any = {};
-    rules: Rules = [];
+    rules: Rules | Rule = [];
     message: string = '';
     text: string = '';
     displayText: string = '';
@@ -72,6 +73,7 @@ export default class Cell extends BaseCell {
     rowExpand = false;
     rowHasChildren = false;
     overflowTooltipShow = true;
+    selectorCellValueType: SelectorCellValueType = 'value';
     overflowTooltipMaxWidth = 500;
     overflowTooltipPlacement: OverflowTooltipPlacement = 'top';
 
@@ -95,6 +97,7 @@ export default class Cell extends BaseCell {
         this.key = column.key;
         this.type = column.type || '';
         this.editorType = column.editorType || 'text';
+        this.selectorCellValueType = column.selectorCellValueType || this.ctx.config.SELECTOR_CELL_VALUE_TYPE || 'value';
         this.editorProps = column.editorProps || {};
         this.cellType = cellType;
         this.align = column.align || 'center';