+++ src/EVirtTable.ts
@@ -160,7 +160,8 @@ export default class EVirtTable {
     }
     loadConfig(_config: ConfigType) {
         this.ctx.config.init(_config);
-        this.ctx.database.init();
+        //重新加载config，初始化表格，但是默认不清除用户操作
+        this.ctx.database.init(false);
         this.header.init();
         // 更新右键菜单，有可能配置项变化
         this.contextMenu.updated();
@@ -199,16 +200,23 @@ export default class EVirtTable {
         this.ctx.off(event, callback);
     }
     filterMethod(func: FilterMethod) {
-        this.scrollTo(0, 0);
         this.ctx.database.setFilterMethod(func);
-        this.ctx.database.init();
+        //重新初始化表格，但是默认不清除用户操作
+        this.ctx.database.init(false);
         this.header.init();
         this.ctx.emit('draw');
     }
     editCell(rowIndex: number, colIndex: number) {
         this.editor.editCell(rowIndex, colIndex);
     }
-    setItemValue(rowKey: string, key: string, value: any, history = true, reDraw = true, isEditor = false) {
+    setItemValue(
+        rowKey: string,
+        key: string,
+        value: any,
+        history = true,
+        reDraw = true,
+        isEditor = false
+    ) {
         this.ctx.database.setItemValue(rowKey, key, value, history, reDraw, isEditor);
     }
     batchSetItemValue(list: ChangeItem[], history = true) {
@@ -217,6 +225,12 @@ export default class EVirtTable {
     setItemValueByEditor(rowKey: string, key: string, value: any, history = true, reDraw = true) {
         this.ctx.setItemValueByEditor(rowKey, key, value, history, reDraw);
         this.editor.doneEdit();
+    }
+
+    clearEditableData(value = null) {
+        const xArr = [0, this.ctx.maxColIndex];
+        const yArr = [0, this.ctx.maxRowIndex];
+        return this.selector.clearSelectedData(xArr, yArr, false, value);
     }
     clearEditor() {
         this.editor.clearEditor();
@@ -299,6 +313,13 @@ export default class EVirtTable {
                     if (Array.isArray(err) && err.length) {
                         const [_err] = err;
                         const { rowKey, key } = _err;
+                        const targetRow = this.ctx.database.getRowForRowKey(rowKey);
+                        if (targetRow) {
+                            const { parentRowKeys = [] } = targetRow;
+                            if (parentRowKeys && parentRowKeys.length) {
+                                this.setExpandRowKeys(parentRowKeys, true);
+                            }
+                        }
                         this.scrollToRowkey(rowKey);
                         this.scrollToColkey(key);
                     }