+++ CHANGE-LOG.md
@@ -218,13 +218,14 @@ fix: 浏览器修复部分缩放canvas字体会糊
 
 V1.2.11
 ### Bug Fixes
-fix: 废除ENABLE_RESERVE_SELECTION属性，默认支持跨页选，如果不想跨页自行清除
+fix: 废除ENABLE_RESERVE_SELECTION属性，默认支持跨页选
 fix: 在handleContextMenu中添加preventDefault以防止默认右键菜单
 feat: 添加列宽度的最小和最大限制
 fix: 修复tree重新加载数据会恢复初始状态
 fix: 修复列宽调整后双击错乱
 feat: 添加获取高亮行功能并更新相关逻辑
 ---
+
 
 V1.2.12
 ### Bug Fixes
@@ -240,4 +241,40 @@ docs: 添加hover编辑图标文档例子
 V1.2.14
 ### Bug Fixes
 fix: 修复dom选中时不能复制
+---
+
+V1.2.15
+### Bug Fixes
+fix: 修复初始化列宽为null时滚动条计算错误
+fix: 优化tooltip功能，添加表头tooltip
+---
+
+V1.2.16
+### Bug Fixes
+fix: 修复loadConfig，filterMethod会清除校验错误信息及已经改变值
+fix: 修复tree校验错误时不回展开
+---
+
+V1.2.17
+### Bug Fixes
+refactor:重构校验器，剔除async-validator
+fix: 添加config类型提示
+---
+
+V1.2.18
+### Bug Fixes
+fix: 解决选择和tree导致的内存泄漏 #73
+feat: 添加选择器值类型支持，更新相关逻辑以处理displayText和value
+---
+
+V1.2.19
+### Bug Fixes
+feat: 优化批量设置值的方法，支持只读设置方法
+feat: 添加清除可以编辑cell数据方法
+---
+
+V1.2.20
+### Bug Fixes
+fix: 废除批量设置值支持只读，变更为通过BEFORE_VALUE_CHANGE_METHOD进行更改，支持只读值更改
+fix: 优化验证逻辑修复非必填时返回错误
 ---